解决跨域问题，并且重构了下数据方面的代码

+++ src/main/java/com/km/service/ProcessModule/service/ProcessService.java
@@ -4,9 +4,11 @@ package com.km.service.ProcessModule.service;
 import com.km.service.ProcessModule.Mapper.ProcessMapper;
 import com.km.service.ProcessModule.domain.Process;
 import com.km.service.ProcessModule.dto.ProcessUseridDto;
+import com.km.service.common.exception.serviceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
@@ -32,6 +34,10 @@ public class ProcessService {
     }
 
     public void deleteProcess(String processId) {
+        Process process = processMapper.getProcessByProcessId(processId);
+        if(process.getRunningJobCount()!=0){
+            throw new serviceException("目前有正在运行的任务绑定着该流程，无法删除");
+        }
         processMapper.deleteProcess(processId);
     }
 
@@ -47,12 +53,21 @@ public class ProcessService {
         return processMapper.getProcessByProcessId(processId);
     }
 
-    public List<ProcessUseridDto> findAllProcess(int pageSize, int pageNumber) {
+    public List<ProcessUseridDto> getAllProcess(int pageSize, int pageNumber) {
         int start = (pageNumber-1)*pageSize;
-        return processMapper.findAllProcess(start,pageSize);
+        return processMapper.getAllProcess(start,pageSize);
     }
 
     public int getProcessCount() {
-        return processMapper.processCount();
+        return processMapper.getProcessCount();
+    }
+
+    public List<Process> getProcessList(String processIds) {
+        List<Process> processes = new ArrayList<>();
+        String[] ids = processIds.split("\t");
+        for(String id:ids){
+            processes.add(processMapper.getProcessByProcessId(id));
+        }
+        return processes;
     }
 }
